Endrer periode til type då useFormState ikke virker sammen med interface extends
