+++ src/frontend/Sider/Behandling/Inngangsvilkår/Aktivitet/LeggTilAktivitet.tsx
@@ -14,6 +14,7 @@ import useFormState, { FormState } from '../../../../hooks/felles/useFormState';
 import { Feilmelding } from '../../../../komponenter/Feil/Feilmelding';
 import DateInput from '../../../../komponenter/Skjema/DateInput';
 import { RessursStatus } from '../../../../typer/ressurs';
+import { Periode } from '../../../../utils/periode';
 import { Aktivitet, AktivitetType } from '../typer';
 
 const Container = styled.div`
@@ -37,10 +38,8 @@ const Knapp = styled(Button)`
 `;
 
 export type NyAktivitet = {
-    fom: string;
-    tom: string;
     type: AktivitetType;
-};
+} & Periode;
 
 const initFormState = { fom: '', tom: '', type: '' };
